+++ geotrek/diving/factories.py
@@ -55,17 +55,13 @@ class DiveFactory(StructureRelatedDefaultFactory):
 
     @factory.post_generation
     def sources(obj, create, extracted=None, **kwargs):
-        if create:
-            if extracted:
-                for source in extracted:
-                    obj.source.add(source)
+        if create and extracted:
+            obj.source.set(extracted)
 
     @factory.post_generation
     def portals(obj, create, extracted=None, **kwargs):
-        if create:
-            if extracted:
-                for portal in extracted:
-                    obj.portal.add(portal)
+        if create and extracted:
+            obj.portal.set(extracted)
 
 
 class DivingManagerFactory(UserFactory):