+++ docs/changelog.rst
@@ -5,6 +5,12 @@ CHANGELOG
 2.60.0+dev
 -------------------
 
+**New features**
+
+- Add Web Links to Trek endpoints in APIv2
+- Add endpoints for Web Links categories in APIv2
+- Ensure APIv2 returns outdoor sites list and outdoor courses list as ordered by localized name
+
 
 2.60.0 (2021-06-25)
 -------------------